Consolidate lane test files and pin the pinned-UI guard
Test steward pass. The machine-selection feature had fractured into four small
files in `components/lanes/` — a folder already carrying 20 test files for 39
sources. `CreateLaneDialogMachines` is folded into `CreateLaneDialog.test.tsx`
(both render the same component), and the LaneCombobox grouping test moves next
to `LaneCombobox.tsx` in `components/terminals/`, where its subject lives; it
was only parked under `lanes/` by a file-ownership constraint while the feature
was being built.

Adds the one gate finding from /quality that had no test: the
`canMutatePinnedProjectUi` change. Every other gate item was already pinned —
name-only repo matches, the git-config parser cases, the structural worktree
walk, offline lane retention, and per-chat routing not rebinding the tab. This
one I made by hand and left uncovered, which is exactly the gap that let dead
code ship in the first place.
